+++ lib/brazil.dart
@@ -1,6 +1,17 @@
 /// Form fields related to Brazil.
+///
+/// - [BrMobileField] for Brazilian Mobile numbers.
+/// - [BrPhoneField] for Brazilian Telephone numbers.
+/// - [CepField] for CEP (Código de Endereçamento Postal) — Brazilian Postal
+///   Code.
+/// - [CnpjField] for CNPJ (Cadastro Nacional de Pessoa Jurídica) — Brazilian
+///   Company's Registered Number.
+/// - [CpfField] for CPF (Cadastro da Pessoa Física) — Brazilian Social Security
+///   Number.
 library brazil;
 
+export 'src/brazil/br_mobile_field.dart';
+export 'src/brazil/br_phone_field.dart';
 export 'src/brazil/cep_field.dart';
 export 'src/brazil/cnpj_field.dart';
 export 'src/brazil/cpf_field.dart';